# lib/api-zod/src/generated/types/fetchVideosResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoItem } from "./videoItem";

export interface FetchVideosResponse {
  videos: VideoItem[];
  total: number;
}