# lib/api-zod/src/generated/types/videoItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoItem {
  videoId: string;
  title: string;
  duration: string;
  views: string;
  likes: string;
  comments: string;
  publishedAt: string;
  transcript: string;
}